Finished chapter 4: basic DNS server
This chapter basically wraps all the underlying packet read/write logic
and puts it into a UDP server that can respond to DNS queries.

Currently it is not very useful since all it does it delegated the
actual DNS resolution to 8.8.8.8. In the next chapter I will implement
actual recursive resolver.